refactor: unnest sidebar components
