+++ app/components/Sidebar/Database/Database.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import DbTables from './DbTables/DbTables';
-import AddDbTable from './DbTables/AddDbTable';
+import DbTables from '../DbTables/DbTables';
+import AddDbTable from '../DbTables/AddDbTable';
 import { connect } from 'react-redux';
-import { deleteDatabase } from '../../../../actions';
-import { getDbTables } from '../../../../data/tables.actions';
+import { deleteDatabase } from '../../../actions';
+import { getDbTables } from '../../../data/tables.actions';
 
 const Database = React.createClass({
   render: function() {
